// src/domain/enums/error-code.enum.ts
export enum ErrorCode {
    USER_ALREADY_EXISTS = 'USER_ALREADY_EXISTS',
    USER_NOT_FOUNDED = 'USER_NOT_FOUNDED',
    PASSWORD_INVALID = 'PASSWORD_INVALID',
    NOT_AUTHENTICATED = 'NOT_AUTHENTICATED',
    ACTION_NOT_ALLOWED = 'ACTION_NOT_ALLOWED',
    BAD_REQUEST = 'BAD_REQUEST',
    NOT_AUTHORIZED = 'NOT_AUTHORIZED',
    INTERNAL_SERVER_ERROR = 'INTERNAL_SERVER_ERROR',
    FIRSTNAME_IS_REQUIRED = 'FIRSTNAME_IS_REQUIRED',
    LASTNAME_IS_REQUIRED = 'LASTNAME_IS_REQUIRED',
    EMAIL_IS_REQUIRED = 'EMAIL_IS_REQUIRED',
    PASSWORD_IS_REQUIRED = 'PASSWORD_IS_REQUIRED',
    ROLE_IS_REQUIRED = 'ROLE_IS_REQUIRED',
    NOT_VALID_EMAIL = 'NOT_VALID_EMAIL',
    PASSWORD_SHOULD_BE_MORE_THAN_8_CHARACTERS = 'PASSWORD_SHOULD_BE_MORE_THAN_8_CHARACTERS',
};